+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/TrackTablePanel.java
@@ -100,7 +100,8 @@ public final class TrackTablePanel extends EntityTablePanel {
 	}
 
 	private BigDecimal getAmountFromUser() {
-		return Dialogs.input(bigDecimalField()
+		return Dialogs.input()
+						.component(bigDecimalField()
 										.nullable(false)
 										.minimumValue(0))
 						.owner(this)