+++ changelog.md
@@ -74,6 +74,7 @@ Codion Change Log
 - FilterTable.Builder.Model added, replacing the tableModel parameter in FilterTable.builder().
 - FileSelectionDialogBuilder.fileFilter() renamed filter().
 - ProgressWorkerDialogBuilder.BuilderFactory added, replacing overloaded methods in Dialogs.
+- InputDialogBuilder.ComponentStep added, replacing overloaded methods in Dialogs.
 ### is.codion.swing.framework.ui
 - ApplicationPreferences.fromString() bug fixed, FRAME_MAXIMIZED_KEY now used.
 - EntityComboBoxPanel.Builder.EditPanel added, replacing the editPanel parameter in EntityComboBoxPanel.builder().